+++ CHANGELOG.md
@@ -4,6 +4,8 @@
   * Improvements to the compiled size of generated messages code with dart2js.
   * Update Intl pubspec to allow newer version of fixnum
   * Allow adjacent literal strings to be used for message names/descriptions.
+  * Provide a better error message for some cases of bad parameters
+    to plural/gender/select messages.
 
 ## 0.12.4+2
   * update analyzer to '<0.27.0'